+++ src/main/java/com/jcloisterzone/event/play/DoubleTurnEvent.java
@@ -1,0 +1,11 @@
+package com.jcloisterzone.event.play;
+
+public class DoubleTurnEvent extends PlayEvent {
+
+    private static final long serialVersionUID = 1L;
+
+    public DoubleTurnEvent(PlayEventMeta metadata) {
+        super(metadata);
+    }
+
+}