+++ changelog.md
@@ -16,6 +16,8 @@
 * fix: City of C. followers can be now redeployed also onto features closed by abbey.
 * fix: when moves from multiple City of C. quarters are available (eg legal move on cloister and road at 
   one time) there was possibility that follower was moved from wrong quarter (eg. from blacksmith to cloister)
+* fix: builder double turn allowed wrongly add follower into City of C. after each turn part if opponent 
+  feature was scored during first part 
 * fix: broken undo for loaded games
 * fix: don't offer AI unsupported expansion in create game panel after "Play again" is used
 * fix: preserve figure order in player panels